# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todo API",
        "description": "This is an open and free to use API for developers to test their client projects.\n        All api calls require you to pass in <br> your API Key with the querystring\n        <b>?apikey=########-####-####-####-############</b>, <br><br>Please\n        register for your own API Key today at https://todoapi.net/register ",
        "version": "1.0"
    },
    "paths": {
...
        "/api/todos/": {
            "delete": {
                "tags": [
                    "Generic Todo"
                ],
                "summary": "Deletes all Todos",
                "operationId": "Todo.removeAll",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "all api calls require the ?apikey querystring.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "completed",
                        "in": "query",
                        "description": "?completed=true if you want to delete only completed",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns message of outcome"
                    }
                }
            }
        },
        "/api/users/{user_id}/todos/": {
            "delete": {
                "tags": [
                    "User Todo"
                ],
                "summary": "Deletes all Todos for the current user.",
                "operationId": "Todo.user.removeAll",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User Id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "all api calls require the ?apikey querystring.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "completed",
                        "in": "query",
                        "description": "?completed=true if you want to delete only completed",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns message of outcome"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/users/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "User login ",
                "operationId": "User.login",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "all api calls require the ?apikey querystring.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass email and password",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "passord": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login request successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "fef6b64d-2223-5dec-83bf-c2dcd468500d"
                                        },
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        },
                                        "enabled": {
                                            "type": "boolean"
                                        },
                                        "admin": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/logout": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "User logout",
                "operationId": "User.logout",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "all api calls require the ?apikey querystring.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logout successful"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/users/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "User registration. There's a max limit of 10 users per developer account.",
                "operationId": "User.register",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "all api calls require the ?apikey querystring.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Submit an email and password",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "name",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "passord": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login request successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "fef6b64d-2223-5dec-83bf-c2dcd468500d"
                                        },
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        },
                                        "enabled": {
                                            "type": "boolean"
                                        },
                                        "admin": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}": {
            "delete": {
                "tags": [
                    "User",
                    "Admin"
                ],
                "summary": "Deletes the current user, or a specific user if current user is admin",
                "operationId": "User.remove",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "all api calls require the ?apikey querystring.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User Id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns status message."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Lists all users, if the current user has admin rights.",
                "operationId": "User.index",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "all api calls require the ?apikey querystring.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns list of all users."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Deletes all users, current user must have admin rights.",
                "operationId": "User.removeAll",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "all api calls require the ?apikey querystring.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "admin",
                        "in": "query",
                        "description": "?admin=true if you want to also delete the admin account (not developer account)",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns status message."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}